smaller Event size, more plotting functionality

+++ ResultProcessor/Event.cc
@@ -1,17 +1,17 @@
 #include "Event.h"
 
 
-double Event::get(const TString &var) const {
-  double val = 0.;
-  std::map<TString,double>::const_iterator it = vars_.find(var);
+float Event::get(const TString &var) const {
+  float val = 0.;
+  std::map<TString,float>::const_iterator it = vars_.find(var);
   if( it != vars_.end() ) val = it->second;
     
   return val;
 }
   
 
-void Event::set(const TString &var, double val) {
-  std::map<TString,double>::iterator it = vars_.find(var);
+void Event::set(const TString &var, float val) {
+  std::map<TString,float>::iterator it = vars_.find(var);
   if( it != vars_.end() ) {
     it->second = val;
   } else {
